requirements.txt
2to3==1.0
aiohttp==3.6.2
amqp==2.5.2

appdirs==1.4.3
argon2-cffi==19.1.0
asgiref==3.2.3
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
beautifulsoup4==4.9.0
billiard==3.6.1.0
bs4==0.0.1
cachetools==4.0.0
celery==4.3.0
certifi==2019.9.11
cffi==1.13.2
channels==2.3.1
chardet==3.0.4
click==7.1.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cssselect==1.1.0
cycler==0.10.0
daphne==2.4.0
distlib==0.3.0
Django==2.2.7
django-cleanup==3.2.0
django-cors-headers==3.2.0
django-elasticsearch-dsl==0.5.1
django-elasticsearch-dsl-drf==0.17.6
django-eventstream==2.6.0
django-filter==2.0.0
django-grip==2.0.2
django-nine==0.2.2
django-postgres-copy==2.4.2
django-rest-passwordreset==1.0.0
django-rest-swagger==2.2.0
django-smtp-ssl==1.0
django-uuidfield==0.5.0
djangorestframework==3.9.4
dnspython==1.16.0
docutils==0.15.2
elasticsearch==6.4.0
elasticsearch-dsl==6.1.0
et-xmlfile==1.0.1
eventlet==0.25.1
fcm-django==0.2.21
filelock==3.0.12
Flask==1.1.2
flower==0.9.4
future==0.18.2
gevent==1.4.0
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
greenlet==0.4.15
gripcontrol==3.3.1
gunicorn==20.0.4
html5lib==1.0.1
httplib2==0.17.0
humanize==0.5.1
hyperlink==19.0.0
idna==2.8
idna-ssl==1.1.0
importlib-metadata==1.1.0
importlib-resources==1.3.1
incremental==17.5.0
inflection==0.3.1
instaloader==4.2.8
ipaddress==1.0.23
itsdangerous==1.1.0
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
jsonfield==2.0.2
kiwisolver==1.1.0
kombu==4.6.3
lxml==4.4.2
MarkupSafe==1.1.1
matplotlib==3.1.1
monotonic==1.5
more-itertools

multidict==4.7.3
ndg-httpsclient==0.5.1
numpy==1.17.4
oauth2client==4.1.3
openapi-codec==1.3.2
openpyxl==3.0.3
pandas==0.25.3
parsel==1.5.2
paypalrestsdk==1.13.1
Pillow==6.2.1
pipenv==2018.11.26
Protego==0.1.16
psycopg2==2.8.4
psycopg2-binary==2.8.4
pubcontrol==2.4.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
PyDispatcher==2.0.5
pyfcm==1.4.5
PyHamcrest==1.9.0
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.2
PyPDF2==1.26.0
PySocks==1.6.8
python-dateutil==2.8.1
python-docx==0.8.10
python-whois==0.7.2
python2==1.2
pytz==2019.3
Quandl==3.4.8
queuelib==1.5.0
redis==3.3.11
reportlab==3.5.42
requests==2.22.0
requests-toolbelt==0.9.1
rsa==4.0
scipy==1.3.0
Scrapy==1.8.0
service-identity==18.1.0
simplejson==3.16.0
six==1.13.0
smartystreets-python-sdk==4.4.0
sounddevice==0.3.13
soupsieve==2.0
sqlparse==0.3.0
statistics==1.0.3.5
stripe==2.40.0
survey
tornado==6.0.4
twilio==6.26.0
Twisted==19.10.0
txaio==18.8.1
typing-extensions==3.7.4.1
uritemplate==3.0.0
urllib3==1.25.7
vine==1.3.0
virtualenv==16.4.3
virtualenv-clone==0.5.3
w3lib==1.21.0
webencodings==0.5.1
Werkzeug==0.16.0
xhtml2pdf==0.2.4
xlrd==1.2.0
XlsxWriter==1.2.7
xlwt==1.3.0
yarl==1.4.2
zipp==0.6.0
zope.interface==4.7.1
